requirements.txt
APScheduler==3.6.3
asn1crypto==0.24.0
backports-abc==0.5
certifi==2021.5.30
configparser==3.5.0b2
cryptography==2.6.1
decorator==4.4.2
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
futures==3.3.0
ipaddress==1.0.17
keyring==17.1.1
keyrings.alt==3.1.1
numpy==1.16.2
picamera==1.13
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4.3
python-telegram-bot==13.1
pytz==2021.1
pyxdg==0.25
PyYAML==3.13
RPi.GPIO==0.7.0
SecretStorage==2.3.1
singledispatch==3.6.2
six==1.12.0
tornado==5.1.1
typing==3.10.0.0
tzlocal==2.1
